generating Linux logs if RA is installed via flatpak
Moreover, I removed the "cd to the main RA directory" as it is not really necessary and might be difficult for some users to do. Plus it might often require root privileges.
In the way I documented, logs are generated in the homedir of the user and there is one less step to be done.